+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased](https://github.com/qutip/QuantumToolbox.jl/tree/main)
 
-- Support different length for to and from on GeneralDimensions. ([#448])
+- Support different length for `to` and `from` on GeneralDimensions. ([#448])
 
 ## [v0.30.0]
 Release date: 2025-04-12